pyproject.toml
[tool.poetry]
name = "scanpay"
version = "0.1.0"
description = "Backend API for scanpay software"
authors = ["Mwapsam <mwapsam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
qrcode = "^7.4.2"
pillow = "^10.4.0"
django-oauth-toolkit = "^2.4.0"
django-environ = "^0.11.2"
django-anymail = "^11.0.1"
mailtrap = "^2.0.1"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
